+++ polymorphic/admin.py
@@ -102,7 +102,10 @@ class PolymorphicParentModelAdmin(admin.ModelAdmin):
 
     add_type_template = None
     add_type_form = PolymorphicModelChoiceForm
-    
+
+    #: The regular expression to filter the primary key in the URL.
+    #: This accepts only numbers as defensive measure against catch-all URLs.
+    #: If your primary key consists of string values, update this regular expression.
     pk_regex = '(\d+)'